# Conflicts:
#	client/package.json
#	server/liveblog/posts/posts.py

+++ server/liveblog/blogs/tasks.py
@@ -20,7 +20,8 @@ from .app_settings import (BLOGLIST_ASSETS, BLOGSLIST_ASSETS_DIR,
 from .embeds import embed, render_bloglist_embed
 from .exceptions import MediaStorageUnsupportedForBlogPublishing
 from .utils import (check_media_storage, get_blog_path,
-                    get_bloglist_path, is_s3_storage_enabled, get_blog)
+                    get_bloglist_path, is_s3_storage_enabled,
+                    get_blog, can_delete_blog)
 
 logger = logging.getLogger('superdesk')
 
@@ -124,11 +125,11 @@ def _publish_blog_embed_on_s3(blog_or_id, theme=None, output=None, safe=True, sa
 
             logger.warning('Media storage not supported for blog "{}"'.format(blog_id))
             # TODO: Add reverse url function.
-            public_url = '{}://{}/embed/{}/{}{}'.format(app.config['URL_PROTOCOL'],
-                                                        app.config['SERVER_NAME'],
-                                                        blog_id,
-                                                        '{}'.format(output_id) if output_id else '',
-                                                        '/theme/{}'.format(theme) if theme else '')
+            public_url = '{}{}/embed/{}/{}{}'.format(app.config['EMBED_PROTOCOL'],
+                                                     app.config['SERVER_NAME'],
+                                                     blog_id,
+                                                     '{}'.format(output_id) if output_id else '',
+                                                     '/theme/{}'.format(theme) if theme else '')
 
         public_urls = blog.get('public_urls', {'output': {}, 'theme': {}})
         updates = {'public_urls': public_urls}
@@ -304,9 +305,22 @@ def remove_deleted_blogs():
 
     for blog in blog_service.find({'blog_status': 'deleted'}):
         if blog['delete_not_before'] <= utcnow():
-            logger.info('Blog "{}" with id "{}" removed as was marked for deletion'.format(blog['title'], blog['_id']))
-
             blog_id = blog['_id']
-            blog_service.on_delete(blog)
-            blog_service.delete(lookup={'_id': blog_id})
-            archive_service.delete(lookup={'blog': ObjectId(blog_id)})
+            if can_delete_blog(blog):
+                try:
+                    blog_service.on_delete(blog)
+                    blog_service.delete(lookup={'_id': blog_id})
+                    archive_service.delete(lookup={'blog': ObjectId(blog_id)})
+
+                    logger.info('Blog "{}" with id "{}" removed as was marked for deletion'.format(
+                        blog['title'], blog['_id']))
+                except Exception as err:
+                    logger.error('There was a problem removing the blog. {}'.format(err))
+            else:
+                # given it's not possible to remove the blog, then we archive it
+                updates = blog.copy()
+                updates['blog_status'] = 'closed'
+                blog_service.update(blog_id, updates, blog)
+
+                logger.info('Blog "{}" with id "{}" archived as it had syndication and active consumers'.format(
+                    blog['title'], blog['_id']))
